% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1_plotly.R
\name{plot1_plotly}
\alias{plot1_plotly}
\title{Title}
\usage{
plot1_plotly(
  class_records,
  average_class_data_by_day = summarize_records(class_records),
  plot3_ymax = 5,
  plot4_ymax = plot3_ymax,
  plot5_ymax = 3,
  x1 = 0,
  x2 = 0.5,
  x3 = 0.5
)
}
\arguments{
\item{class_records}{}

\item{average_class_data_by_day}{}

\item{plot3_ymax}{}

\item{plot4_ymax}{}

\item{plot5_ymax}{}

\item{x1}{}

\item{x2}{}

\item{x3}{}
}
\description{
Title
}
