% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the agent-based model}
\usage{
run_simulation(
  n_schools = 1000,
  n_grades_per_school = 7,
  n_classes_per_school = n_grades_per_school * ceiling(60/n_students_per_class),
  n_students_per_class = 15,
  n_weeks_in_time_window = 4 * 2,
  pr_student_infected_per_infectious_household_adult_per_day = p_infect_per_day(p_infect
    = 0.16, n_days_infectious = n_days_infectious),
  pr_household_adult_infected_by_infectious_student_per_day = p_infect_per_day(p_infect =
    0.4, n_days_infectious = n_days_infectious),
  pr_household_adult_infected_per_infectious_household_adult_per_day =
    p_infect_per_day(p_infect = 0.4, n_days_infectious = n_days_infectious),
  pr_student_symptomatic_if_infected = 0.5,
  pr_adult_symptomatic_if_infected = 0.7,
  test_specificity = 0.999,
  test_sensitivity_day_0 = 1 - test_specificity,
  test_sensitivity_day_1 = 1 - test_specificity,
  `test sensitivity day 2` = 1 - test_specificity,
  `test sensitivity day 3` = 0.5,
  `test sensitivity day 4` = 0.7,
  `peak test sensitivity` = 0.95,
  `days of peak test sensitivity` = days(3),
  `test sensitivity on last day of symptoms` = 0.5,
  `daily decrease in sensitivity from peak until recovery` =
    (`test sensitivity on last day of symptoms` - `peak test sensitivity`) *
    (days(1)/(last_active_infection_day - (days(4) + `days of peak test sensitivity`))),
  `decrease in sensitivity per day after recovery` = 0.1,
  attestation_sensitivity_symptomatic = 0.9,
  attestation_sensitivity_asymptomatic = max(0.1, 1 - attestation_specificity),
  attestation_sensitivity_presymptomatic = attestation_sensitivity_asymptomatic,
  attestation_specificity = 0.999,
  baseline_prevalence_of_covid_safety_education = 0.1,
  pr_become_educated_after_outreach = 0.5,
  baseline_prevalence_adults = 0.006,
  baseline_prevalence_students = baseline_prevalence_adults,
  baseline_cumulative_incidence_adults = 0.04,
  baseline_cumulative_incidence_students = baseline_cumulative_incidence_adults,
  baseline_recovered_adults = baseline_cumulative_incidence_adults -
    baseline_prevalence_adults,
  baseline_recovered_students = baseline_cumulative_incidence_students -
    baseline_prevalence_students,
  baseline_vaccinated_adults = 0,
  baseline_vaccinated_students = 0,
  increase_in_accuracy_after_education = 0.1,
  decrease_in_exogenous_risk_after_education = 0.1,
  adult_exogenous_infection_risk_per_day_without_education = 4e-04,
  adult_exogenous_infection_risk_per_day_with_education =
    adult_exogenous_infection_risk_per_day_without_education * (1 -
    decrease_in_exogenous_risk_after_education),
  student_exogenous_infection_risk_per_day_without_education =
    adult_exogenous_infection_risk_per_day_without_education,
  student_exogenous_infection_risk_per_day_with_education =
    student_exogenous_infection_risk_per_day_without_education * (1 -
    decrease_in_exogenous_risk_after_education),
  risk_per_infected_close_contact = 0.001,
  risk_per_infected_classmate = 5e-04,
  testing_fraction = 0.25,
  first_infectious_day = days(3),
  last_infectious_day = days(12),
  n_days_infectious = (last_infectious_day - first_infectious_day)/days(1) + 1,
  first_symptomatic_day = days(5),
  last_symptomatic_day = days(14),
  last_active_infection_day = days(14),
  quarantine_length_after_adult_positive_test = days(14),
  quarantine_length_after_student_positive_test = days(14),
  quarantine_length_after_positive_attestation = days(3),
  quarantine_length_after_outbreak = days(14),
  in_person_school_start_date = as.Date("2021-01-04"),
  wait_time_for_surveillance_test_results = days(2),
  wait_time_for_attestation_triggered_test_results = days(2),
  wait_time_for_household_education_after_positive_attestation = days(1),
  wait_time_for_testing_after_outreach = days(1),
  surveillance_testing_days = "Monday",
  days_from_baseline_collection_until_baseline_return = days(2),
  baseline_return_date = baseline_collection_date +
    days_from_baseline_collection_until_baseline_return,
  baseline_collection_date = as.Date("2020-12-05"),
  n_students_recently_in_class_and_diagnosed_for_outbreak = 3,
  cluster_time_window = days(14),
  n_adults_per_household = 2,
  browse_at_start_of_days = FALSE,
  browse_before_return = FALSE,
  n_close_contacts_in_class = 5,
  print_args = FALSE,
  verbose = FALSE,
  plot_data = verbose | browse_at_start_of_days,
  plot3_ymax = 6,
  in_shiny = FALSE
)
}
\arguments{
\item{n_schools}{}

\item{n_grades_per_school}{}

\item{n_classes_per_school}{}

\item{n_students_per_class}{}

\item{n_weeks_in_time_window}{}

\item{pr_student_infected_per_infectious_household_adult_per_day}{}

\item{pr_household_adult_infected_by_infectious_student_per_day}{}

\item{pr_household_adult_infected_per_infectious_household_adult_per_day}{}

\item{pr_student_symptomatic_if_infected}{}

\item{pr_adult_symptomatic_if_infected}{}

\item{test_specificity}{}

\item{test_sensitivity_day_0}{}

\item{test_sensitivity_day_1}{}

\item{attestation_sensitivity_symptomatic}{}

\item{attestation_sensitivity_asymptomatic}{}

\item{attestation_sensitivity_presymptomatic}{}

\item{attestation_specificity}{}

\item{baseline_prevalence_of_covid_safety_education}{}

\item{pr_become_educated_after_outreach}{}

\item{baseline_prevalence_adults}{}

\item{baseline_prevalence_students}{}

\item{baseline_cumulative_incidence_adults}{}

\item{baseline_cumulative_incidence_students}{}

\item{baseline_recovered_adults}{}

\item{baseline_recovered_students}{}

\item{baseline_vaccinated_adults}{}

\item{baseline_vaccinated_students}{}

\item{increase_in_accuracy_after_education}{}

\item{decrease_in_exogenous_risk_after_education}{}

\item{adult_exogenous_infection_risk_per_day_without_education}{}

\item{adult_exogenous_infection_risk_per_day_with_education}{}

\item{student_exogenous_infection_risk_per_day_without_education}{}

\item{student_exogenous_infection_risk_per_day_with_education}{}

\item{risk_per_infected_close_contact}{}

\item{risk_per_infected_classmate}{}

\item{testing_fraction}{}

\item{first_infectious_day}{}

\item{last_infectious_day}{}

\item{n_days_infectious}{}

\item{first_symptomatic_day}{}

\item{last_symptomatic_day}{}

\item{last_active_infection_day}{}

\item{quarantine_length_after_adult_positive_test}{}

\item{quarantine_length_after_student_positive_test}{}

\item{quarantine_length_after_positive_attestation}{}

\item{quarantine_length_after_outbreak}{}

\item{in_person_school_start_date}{}

\item{wait_time_for_surveillance_test_results}{}

\item{wait_time_for_attestation_triggered_test_results}{}

\item{wait_time_for_household_education_after_positive_attestation}{}

\item{wait_time_for_testing_after_outreach}{}

\item{surveillance_testing_days}{}

\item{days_from_baseline_collection_until_baseline_return}{}

\item{baseline_return_date}{}

\item{baseline_collection_date}{}

\item{n_students_recently_in_class_and_diagnosed_for_outbreak}{}

\item{cluster_time_window}{}

\item{n_adults_per_household}{}

\item{browse_at_start_of_days}{}

\item{browse_before_return}{}

\item{n_close_contacts_in_class}{}

\item{print_args}{}

\item{verbose}{}

\item{plot_data}{}

\item{plot3_ymax}{}

\item{in_shiny}{used to provide progress reporting to the dashboard when running the model through the Shiny GUI}

\item{`peak}{test sensitivity` The maximum test sensitivity achieved during an infection}

\item{`days}{of peak test sensitivity`}

\item{`test}{sensitivity on last day of symptoms`}

\item{`daily}{decrease in sensitivity from peak until recovery`}

\item{`decrease}{in sensitivity per day after recovery`}
}
\description{
Run the agent-based model
}
