% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_class_records.R
\name{summarize_records}
\alias{summarize_records}
\title{Title}
\usage{
summarize_records(
  class_records,
  label_columns = c("date", "day_of_week", "school_day", "surveillance testing today",
    "school", "class"),
  summary_grouping_columns = c("date", "school_day"),
  school_records_by_date = class_records \%>\% group_by(date, school_day, school) \%>\%
    summarize(.groups = "drop", n_classes_in_school = n(), n_classes_quarantined_today =
    sum(class_quarantined_today), across(any_of(c("n_students_quarantined_today",
    "n households associated with class",
    "n households with positive attestations today",
    "n households that have received outreach",
    "n_household_adults_associated_with_class",
    "n household adults educated about COVID safety",
    "n household adults newly infected today",
    "n household adults infected from exogenous source today",     
    "n household adults infected by student today",
    "n household adults infected (cumulative)",
    "n household adults infected from exogenous sources (cumulative)",
    "n household adults infected since baseline (cumulative)",
    "n household adults infected by student (cumulative)",
    "n household adults infected by other household adult (cumulative)",
    "n_symptomatic_students_in_classroom_today", "n_students_in_attendance_today",
    "n_students_on_roster", "n students newly infected today",
    "n students educated about COVID safety (cumulative)",     
    "n_students_infected_cumulative", "n_students_infected_since_baseline_cumulative",
    "n students infected from exogenous sources (cumulative)",
    "n students infected by household adult (cumulative)",
    "n_students_infected_outside_school_cumulative",
    "n_transmissions_in_class_cumulative", "n_outbreaks",
    "n_initially_uninfected_students_in_classroom_today",
    "n_transmissions_in_class_today",
    "n students newly infected from exogenous source today",
    "n_infectious_students_in_classroom_today", "n_students_at_home_today",     
    "n_uninfectious_students_at_home_today", "n_infectious_students_at_home_today",
    "n_initially_infected_students_in_classroom_today")), sum))
)
}
\arguments{
\item{class_records}{}

\item{label_columns}{}

\item{summary_grouping_columns}{}

\item{school_records_by_date}{}
}
\value{

}
\description{
Title
}
